api/requirements.txt
flask>=3.0
boto3>=1.34
python-dotenv>=1.0
werkzeug>=3.0
